feat:自由对话任务详情按钮常驻;perf:自由对话结束后只清空todolist数据
